# internal/api/docs/swagger.yaml
definitions:
...
  api.GetFeedOutput:
    properties:
      feed:
        $ref: '#/definitions/db.Feed'
    type: object
  api.GetFeedPurchaseOutput:
    properties:
      feed_purchase:
        $ref: '#/definitions/db.FeedPurchase'
    type: object
  api.GetFeedPurchasesOutput:
    properties:
      feed_purchases:
        items:
          $ref: '#/definitions/db.FeedPurchase'
        type: array
    type: object
  api.GetFeedsOutput:
    properties:
      feeds:
        items:
          $ref: '#/definitions/db.Feed'
        type: array
    type: object
...
  api.UpsertFeedInput:
    properties:
      name:
        type: string
      projectid:
        type: string
    required:
    - name
    - projectid
    type: object
  api.UpsertFeedPurchaseInput:
    properties:
      amount_purchased:
        type: number
      date_purchased:
        type: string
      feedid:
        type: string
      projectid:
        type: string
      total_cost:
        type: number
    required:
    - amount_purchased
    - date_purchased
    - feedid
    - projectid
    - total_cost
    type: object
...
  db.Feed:
    properties:
      created:
        type: string
      id:
        type: string
      name:
        type: string
      projectid:
        type: string
      updated:
        type: string
      userid:
        type: string
    type: object
  db.FeedPurchase:
    properties:
      amount_purchased:
        type: number
      created:
        type: string
      date_purchased:
        type: string
      feedid:
        type: string
      id:
        type: string
      projectid:
        type: string
      total_cost:
        type: number
      updated:
        type: string
      userid:
        type: string
    type: object
...
info:
  contact: {}
  title: 4H Record Books API
paths:
...
  /feed:
    get:
      consumes:
      - application/json
      description: Gets all of a user's feeds given a project ID
      parameters:
      - description: Project ID
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetFeedsOutput'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get feeds by project
      tags:
      - Feed
    post:
      consumes:
      - application/json
      description: Adds a feed to a user's personal records
      parameters:
      - description: Feed information
        in: body
        name: UpsertFeedInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertFeedInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Add a feed
      tags:
      - Feed
  /feed-purchase:
    get:
      consumes:
      - application/json
      description: Gets all of a user's feed purchases given a project ID
      parameters:
      - description: Project ID
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetFeedPurchasesOutput'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get feed purchases by project
      tags:
      - Feed Purchase
    post:
      consumes:
      - application/json
      description: Adds a feed purchase to a user's personal records
      parameters:
      - description: Feed Purchase information
        in: body
        name: UpsertFeedPurchaseInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertFeedPurchaseInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Add a feed purchase
      tags:
      - Feed Purchase
  /feed-purchase/{feedPurchaseId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's feed purchase given the feed purchase ID
      parameters:
      - description: Feed Purchase ID
        in: path
        name: feedPurchaseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes a feed purchase
      tags:
      - Feed Purchase
    get:
      consumes:
      - application/json
      description: Get a user's feed purchase by ID
      parameters:
      - description: Feed Purchase ID
        in: path
        name: feedPurchaseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetFeedPurchaseOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get a feed purchase
      tags:
      - Feed Purchase
    put:
      consumes:
      - application/json
      description: Updates a user's feed purchase information
      parameters:
      - description: Feed Purchase ID
        in: path
        name: feedPurchaseId
        required: true
        type: string
      - description: Feed purchase information
        in: body
        name: UpsertFeedPurchaseInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertFeedPurchaseInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update a feed purchase
      tags:
      - Feed Purchase
  /feed/{feedId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's feed given the feed ID
      parameters:
      - description: Feed ID
        in: path
        name: feedId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes a feed
      tags:
      - Feed
    get:
      consumes:
      - application/json
      description: Get a user's feed by ID
      parameters:
      - description: Feed ID
        in: path
        name: feedId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetFeedOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get a feed
      tags:
      - Feed
    put:
      consumes:
      - application/json
      description: Updates a user's feed information
      parameters:
      - description: Feed ID
        in: path
        name: feedId
        required: true
        type: string
      - description: Feed information
        in: body
        name: UpsertFeedInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertFeedInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update a feed
      tags:
      - Feed
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"